requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
jmd_imagescraper>=0.3.0
faiss-cpu>=1.7.0